refactor(library): move registration into library::registry
Registration exists to produce side effects (dummy trees, tags.json),
so unlike bundling there is no pure core to protect; the library layer
owns the IO wholesale. commands/library.rs shrinks to input validation,
display formatting, and final stdout messages, and commands/bundle.rs
stops reaching into a sibling command handler. STD_ID now has a single
definition in the registry. Shared test fixtures live in
library::testutil.
