+++ src/library/mod.rs
@@ -1,9 +1,14 @@
 //! 登録済みライブラリのドメイン。`tags.json` のデータモデルと永続化、`$LOCAL` 上のストレージ配置、
-//! 登録時の識別子抽出・ダミー生成・集約ハッシュ計算をまとめる。`library` / `bundle` の両コマンドが
-//! 共有する中核で、走査の詳細は [`crate::fs`] に委ねる。
+//! 識別子抽出・ダミー生成・集約ハッシュ計算と、それらを束ねる登録処理 (registry) をまとめる。
+//! `library` / `bundle` の両コマンドが共有する中核で、走査の詳細は [`crate::fs`] に、コンパイラへの
+//! 問い合わせは [`crate::compiler`] に委ねる。
 
 pub mod dummy;
 pub mod hash;
 pub mod identifiers;
 pub mod local;
+pub mod registry;
 pub mod tags;
+
+#[cfg(test)]
+pub mod testutil;
